+++ src/main/java/com/lmax/disruptor/RunState.java
@@ -1,8 +1,0 @@
-package com.lmax.disruptor;
-
-enum RunState
-{
-    IDLE,
-    RUNNING,
-    HALTED
-}